add check_number_threads to ot/lp/__init__.py __all__
